better logging and checking when cloning tf src code repo
Sometimes the terraform source code repo is not cloned correctly. When
this happens the InvalidGitRepositoryError exception is raised. To fix
it the user must delete the directory and try again but they did't know
the path to the directory. This commit logs the path to the directory
but also catches the exception and removes the directory and reclones it.