fix and refactor terraform documentation section parsing
The previous logic was very brittle and would fail to parse a section if
the section contained any lines that were not valid attributes or white
space. This commit fixes that by simply ignoring lines that are not
valid attributes and keeping the rest. It will debug log the lines that
are ignored.

While I was in here messing around I went ahead and use Co-Pilot to
generate some of the missing pydocs for the doc_file module.