Publish release canidates to the real PyPi repository
Previously we were publishing release candidates to the test PyPi, but
this made is harder for users to test the release candidates.

I also made changes to the job so that it will work for users on forked
repositories. This uses instructions I got from @mheap blog on checking
the permissions of the underlying user.