Fix bug where GetAttr was not being resolved correctly
There was a bug where GetAttr only worked when it was nested inside of
another instrinsic function. This is because the get_attr function
expected a list of strings with a length of 2. This worked fine when
nested inside of !Sub, because it was converted to a list of strings. But
when it was not nested, it was passed as a string like "!GetAttr foo.bar".

I also fixed an issue with how failed instrinsic functions were logged.
Previously they were skipped and a comment was left inside the converted
terraform. The problem was the error msg was omitted so it wasn't clear
what went wrong. Now the comment includes the error message.

Fixes #243