change the devcontainer remote user to ubuntu
The new ubuntu 24.x image includes a user named `ubuntu` that has a UID
of 1000. I tried to create a vscode user with UID of 1001 but it didn't
work. I tried several things but eventually I gave up and embraced the
`ubuntu` user.

I had to delete the .nox directory to get pre-commit to work again. I
also needed to manually select the python interpreter in vscode.